refactor(tui): centralize plan editor handling
